+++ CloudSchoolBusParent/etonkids-iphone/Class/Health/ETPraiseViewController.h
@@ -15,7 +15,7 @@
 {
     UIView * topView;
     UIView *bottomView;
-    int star;
+    NSInteger star;
     
     MBProgressHUD *HUD;
     UILabel *startlabl;
